requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
# gspread_dataframe 
pandas

# binance API
python-binance
python-dotenv
requests